+++ train_setup/env_wrapper.py
@@ -6,14 +6,14 @@ class EnvWrapper(gym.Wrapper):
         gym.Wrapper.__init__(self, env)
     def reset(self):
         state = self.env.reset()[0]
-        return torch.from_numpy(state).unsqueeze_(dim=0).double()
+        return torch.from_numpy(state).unsqueeze_(dim=0)
 
     def step(self, action):
         action = action.squeeze(0).numpy()
         next_state, reward, terminated, truncated, info = self.env.step(action)
-        next_state = torch.from_numpy(next_state).unsqueeze_(dim=0).double()
-        reward = torch.tensor(reward).view(1, -1).double()
-        terminated = torch.tensor(terminated).view(1, -1).double()
-        truncated = torch.tensor(truncated).view(1, -1).double()
+        next_state = torch.from_numpy(next_state).unsqueeze_(dim=0)
+        reward = torch.tensor(reward).view(1, -1)
+        terminated = torch.tensor(terminated).view(1, -1)
+        truncated = torch.tensor(truncated).view(1, -1)
 
         return next_state, reward, terminated, truncated, info
